refactor: extract `eventDetailSectionHeader`
